passaralhos-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]